+++ tests/testthat/test_silent_messages.R
@@ -11,8 +11,8 @@ context("Silent messages")
 
 # set up test context
 options("tryCatchLog.dump.errors.to.file" = FALSE)
-options("tryCatchLog.silent.warnings" = FALSE)
-options("tryCatchLog.silent.messages" = FALSE)
+options("tryCatchLog.silent.warnings"     = FALSE)
+options("tryCatchLog.silent.messages"     = FALSE)